+++ scripts/device/ROS_dome.py
@@ -1,60 +1,45 @@
 #!/usr/bin/env python3
 
 import time
-import math
 import threading
-import pyinterface
 import sys
+sys.path.append('/home/amigos/ros/src/necst/lib')
 sys.path.append('/home/necst/ros/src/necst/lib')
 import dome_pos
-
+import dome_device
 #ROS
 sys.path.append("/opt/ros/kinetic/lib/python2.7/dist-packages")
 import rospy
 from necst.msg import Status_encoder_msg
 from necst.msg import Status_dome_msg
 from necst.msg import Dome_msg
+from necst.msg import Bool_necst
 
 node_name = "dome"
+import topic_status
+deco = topic_status.deco(node_name)
 
 class dome_controller(object):
     touchsensor_pos = [-391,-586,-780,-974,-1168,-1363,-1561,-1755,-1948,-2143, 0, -197]
     dome_encoffset = 10000
-    buffer = [0,0,0,0,0,0]
-    stop = [0]
-    error = []
     count = 0
-    dome_enc = 0
     limit = 0
-
-    ###status_paramter
-    move_status = ''
-    right_act = ''
-    right_pos = ''
-    left_act = ''
-    left_pos = ''
-    memb_act = ''
-    memb_pos = ''
-    remote_status = ''
+    dome_enc = 0
 
     ###ROS_dome.py parameter
     enc_az = 0###antenna az for dome tracking
 
     parameters = {
-        'target_az':0,
         'command':0
         }
-    
+    paralist = []
+    parameter_az = 0
     ###command flags
-    flag = 0
-
     end_flag = True
-           
+
     def __init__(self):
-        board_name = 2724
-        rsw_id = 2
         self.dome_pos = dome_pos.dome_pos_controller()
-        self.dio = pyinterface.open(board_name, rsw_id)
+        self.dev = dome_device.dome_device()
         self.read_init_domepos()
         self.start_status_check()
         pass
@@ -67,6 +52,7 @@ class dome_controller(object):
         except:
             pass
         
+
     def start_thread(self):
         tlist = threading.enumerate()
         for i in tlist:
@@ -74,7 +60,7 @@ class dome_controller(object):
             if n == "d_track":
                 return
         self.end_track_flag = threading.Event()
-        self.thread = threading.Thread(target = self.move_track, name = "d_track")
+        self.thread = threading.Thread(target = self.dev.move_track, name = "d_track")
         self.thread.start()
         return
     
@@ -83,327 +69,63 @@ class dome_controller(object):
             self.end_track_flag.set()
             self.thread.join()
             buff = [0]
-            self.dio.output_point(buff, 2)
+            self.dev.dio.output_point(buff, 2)
             return
         except:
             pass
-    
-    def move_track(self):
-        print('dome_trakcking start', self.end_flag)
-        ret = self.read_domepos()
-        while not self.end_flag:
-            enc_az = float(self.enc_az)
-            dome_az = self.read_domepos()
-            dome_az = dome_az/3600.
-            self.dome_limit()
-            enc_az = float(enc_az)
-            enc_az = enc_az/3600.
-            if math.fabs(enc_az - dome_az) >= 2.0:
-                self.move(enc_az, track=True)
-            time.sleep(0.01)
-            if self.end_flag == True:
-                break
-            print('dome_tracking')
-    
+
     def test(self, num): #for track_test
         self.start_thread()
         time.sleep(num)
         self.end_thread()
         return
-    
-    def print_msg(self,msg):
-        print(msg)
-        return
-    
-    def print_error(self,msg):
-        self.error.append(msg)
-        self.print_msg('!!!!ERROR!!!!'+msg)
-        return
-    
+
     def move_org(self):
         dist = 90
-        self.move(dist)    #move_org
+        pos = self.dome_enc
+        dir = self.dev.move(dist, pos)#move_org
+        while not dir <= 0.5:
+            pos = self.dome_enc
+            dir = self.dev.move(dist, pos)
         return
-    
-    def move(self, dist, track=False):
-        pos_arcsec = float(self.dome_enc)#[arcsec]
-        pos = pos_arcsec/3600.
-        pos = pos % 360.0
-        dist = float(dist) % 360.0
-        diff = dist - pos
-        dir = diff % 360.0
-        if dir < 0:
-            dir = dir*(-1)
-        
-        if pos == dist: return
-        if dir < 0:
-            if abs(dir) >= 180:
-                turn = 'right'
-            else:
-                turn = 'left'
-        else:
-            if abs(dir) >= 180:
-                turn = 'left'
-            else:
-                turn = 'right'
-        if abs(dir) < 5.0 or abs(dir) > 355.0 :
-            speed = 'low'
-        elif abs(dir) > 15.0 and abs(dir) < 345.0:###or => and
-            speed = 'high'
-        else:
-            speed = 'mid'
-        if not abs(dir) < 0.5:
-            global buffer
-            self.buffer[1] = 1
-            self.do_output(turn, speed)
-            if track:
-                return
-            while dir != 0:
-                pos_arcsec = float(self.dome_enc)
-                pos = pos_arcsec/3600.
-                pos = pos % 360.0
-                dist = dist % 360.0
-                diff = dist - pos
-                dir = diff % 360.0
-                if abs(dir) <= 0.5:
-                    dir = 0
-                else:
-                    if dir < 0:
-                        if abs(dir) >= 180:
-                            turn = 'right'
-                        else:
-                            turn = 'left'
-                    else:
-                        if abs(dir) >= 180:
-                            turn = 'left'
-                        else:
-                            turn = 'right'
-                            
-                    if abs(dir) < 5.0 or abs(dir) > 355.0:
-                        speed = 'low'
-                    elif abs(dir) > 20.0 and abs(dir) < 340.0:###or => and
-                        speed = 'high'
-                    else:
-                        speed = 'mid'
-                    self.do_output(turn, speed)
-                time.sleep(0.1)
-        
-        self.dome_stop()
+
+    def con_move_track(self):
+        print("dome_tracking start", self.end_flag)
+        while not self.end_flag:
+            dome_az = self.dome_enc
+            self.dev.move_track(self.enc_az, dome_az)
+        if self.end_flag == True:
+            while "dome_tracking" in self.paralist:
+                self.paralist.remove("dome_tracking")
         return
-    
-    def dome_stop(self):
-        buff = [0]
-        self.dio.output_point(buff, 2)
-        self.flag = 0
-        return
-    
-    def dome_open(self):
-        ret = self.get_door_status()
-        if ret[1] != 'OPEN' and ret[3] != 'OPEN':
-            buff = [1, 1]
-            self.dio.output_point(buff, 5)
-            d_door = self.get_door_status()
-            while ret[1] != 'OPEN':
-                time.sleep(5)
-                ret = self.get_door_status()
-        buff = [0, 0]
-        self.dio.output_point(buff, 5)
-        return
-    
-    def dome_close(self):
-        ret = self.get_door_status()
-        if ret[1] != 'CLOSE' and ret[3] != 'CLOSE':
-            buff = [0, 1]
-            self.dio.output_point(buff, 5)
-            while ret[1] != 'CLOSE':
-                time.sleep(5)
-                ret = self.get_door_status()
-        buff = [0, 0]
-        self.dio.output_point(buff, 5)
-        return
-    
-    def memb_open(self):
-        ret = self.get_memb_status()
-        if ret[1] != 'OPEN':
-            buff = [1, 1]
-            self.dio.output_point(buff, 7)
-            while ret[1] != 'OPEN':
-                time.sleep(5)
-                ret = self.get_memb_status()
-        buff = [0, 0]
-        self.dio.output_point(buff, 7)
-        return
-    
-    def memb_close(self):
-        ret = self.get_memb_status()
-        if ret[1] != 'CLOSE':
-            buff = [0, 1]
-            self.dio.output_point(buff, 7)
-            while ret[1] != 'CLOSE':
-                time.sleep(5)
-                ret = self.get_memb_status()
-        buff = [0, 0]
-        self.dio.output_point(buff, 7)
-        return
-    
-    """
-    def emergency_stop(self):
-        global stop
-        dome_controller.stop = [1]
-        #self.pos.dio.do_output(self.stop, 11, 1)
-        self.print_msg('!!EMERGENCY STOP!!')
-        return
-    """
-    
-    def dome_fan(self, fan):
-        if fan == 'on':
-            fan_bit = [1, 1]
-            self.dio.output_point(fan_bit, 9)
-        else:
-            fan_bit = [0, 0]
-            self.dio.output_point(fanbit, 9)
+
+    def con_move(self, dist):
+        while not self.end_flag:
+            pos = self.dome_enc
+            dir = self.dev.move(dist, pos)
+            if dir <= 0.5:
+                self.paralist.remove("dome_move")
+                self.dev.dome_stop()
+                break
+        if self.end_flag:
+            while "dome_move" in self.paralist:
+                self.paralist.remove("dome_move")
         return
     
     def get_count(self):
         self.count = self.dome_pos.dome_encoder_acq()
-        return self.count
-    
-    def do_output(self, turn, speed):
-        global buffer
-        global stop
-        if turn == 'right': self.buffer[0] = 0
-        else: self.buffer[0] = 1
-        if speed == 'low':
-            self.buffer[2:4] = [0, 0]
-        elif speed == 'mid':
-            self.buffer[2:4] = [1, 0]
-        else:
-            self.buffer[2:4] = [0, 1]
-        if dome_controller.stop[0] == 1:
-            self.buffer[1] = 0
-        else: self.buffer[1] = 1
-        self.dio.output_point(self.buffer, 1)
-        print('do_output')
-        return
-    
-    def get_action(self):
-        ret = self.dio.input_point(1, 1)
-        if ret == 0:
-            self.move_status = 'OFF'
-        else:
-            self.move_status = 'DRIVE'
-        return
-    
-    def get_door_status(self):
-        ret = self.dio.input_point(2, 6)
-        if ret[0] == 0:
-            self.right_act = 'OFF'
-        else:
-            self.right_act = 'DRIVE'
-        
-        if ret[1] == 0:
-            if ret[2] == 0:
-                self.right_pos = 'MOVE'
-            else:
-                self.right_pos = 'CLOSE'
-        else:
-            self.right_pos = 'OPEN'
-        
-        if ret[3] == 0:
-            self.left_act = 'OFF'
-        else:
-            self.left_act = 'DRIVE'
-        
-        if ret[4] == 0:
-            if ret[5] == 0:
-                self.left_pos = 'MOVE'
-            else:
-                self.left_pos = 'CLOSE'
-        else:
-            self.left_pos = 'OPEN'
-        return
-        
-    def get_memb_status(self):
-        ret = self.dio.input_point(8, 3)
-        if ret[0] == 0:
-            self.memb_act = 'OFF'
-        else:
-            self.memb_act = 'DRIVE'
-        
-        if ret[1] == 0:
-            if ret[2] == 0:
-                self.memb_pos = 'MOVE'
-            else:
-                self.memb_pos = 'CLOSE'
-        else:
-            self.memb_pos = 'OPEN'
-        return
-    
-    def get_remote_status(self):
-        ret = self.dio.input_point(11, 1)
-        if ret[0] == 0:
-            self.remote_status = 'REMOTE'
-        else:
-            self.remote_status = 'LOCAL'
-        return status
-    
-    def error_check(self):
-        ret = self.dio.input_point(16, 6)
-        if ret[0] == 1:
-            self.print_error('controll board sequencer error')
-        if ret[1] == 1:
-            self.print_error('controll board inverter error')
-        if ret[2] == 1:
-            self.print_error('controll board thermal error')
-        if ret[3] == 1:
-            self.print_error('controll board communication error')
-        if ret[4] == 1:
-            self.print_error('controll board sequencer(of dome_door or membrane) error')
-        if ret[5] == 1:
-            self.print_error('controll board inverter(of dome_door or membrane) error')
         return
     
     def limit_check(self):
         while True:
-            limit1 = self._limit_check()
+            limit1 = self.dev._limit_check()
             time.sleep(0.002)
-            limit2 = self._limit_check()
+            limit2 = self.dev._limit_check()
             if limit1 == limit2:
                 return limit1
             continue
         pass
-        
-    def _limit_check(self):
-        limit = self.dio.input_point(12, 4)
-        ret = 0
-        if limit[0:4] == [0,0,0,0]:
-            ret = 0
-        elif limit[0:4] == [1,0,0,0]:
-            ret = 1
-        elif limit[0:4] == [0,1,0,0]:
-            ret = 2
-        elif limit[0:4] == [1,1,0,0]:
-            ret = 3
-        elif limit[0:4] == [0,0,1,0]:
-            ret = 4
-        elif limit[0:4] == [1,0,1,0]:
-            ret = 5
-        elif limit[0:4] == [0,1,1,0]:
-            ret = 6
-        elif limit[0:4] == [1,1,1,0]:
-            ret = 7
-        elif limit[0:4] == [0,0,0,1]:
-            ret = 8
-        elif limit[0:4] == [1,0,0,1]:
-            ret = 9
-        elif limit[0:4] == [0,1,0,1]:
-            ret = 10
-        elif limit[0:4] == [1,1,0,1]:
-            ret = 11
-        elif limit[0:4] == [0,0,1,1]:
-            ret = 12
-        return ret
-    
+
     def dome_limit(self):
         limit = self.limit_check()
         if limit != 0:
@@ -421,9 +143,6 @@ class dome_controller(object):
         f.close()
         return 
     
-    def read_limit(self):
-        return self.limit
-    
     def start_status_check(self):
         th1 = threading.Thread(target = self.status_check)
         th1.setDaemon(True)
@@ -432,26 +151,12 @@ class dome_controller(object):
     
     def status_check(self):
         while not rospy.is_shutdown():
-            self.get_action()
-            self.get_door_status()
-            self.get_memb_status()
-            self.get_remote_status()
+            self.dev.get_action()
+            self.dev.get_door_status()
+            self.dev.get_memb_status()
+            self.dev.get_remote_status()
             self.get_domepos()
             time.sleep(0.1)
-    
-    def stop_status_check(self):
-        self.stop_staus_flag.set()
-        self.status_thread.join()
-        return
-    
-    def read_count(self):
-        return self.count
-    
-    def read_status(self):
-        return self.status_box
-    
-    def read_domepos(self):
-        return self.dome_enc
 
     ###ROS part
 
@@ -460,12 +165,20 @@ class dome_controller(object):
         th = threading.Thread(target = self.pub_status)
         th.setDaemon(True)
         th.start()
-        th2 = threading.Thread(target = self.act_dome)
+        th2 = threading.Thread(target = self.dome_OC)
         th2.setDaemon(True)
         th2.start()
-        th3 = threading.Thread(target = self.stop_dome)
+        th3 = threading.Thread(target = self.memb_OC)
         th3.setDaemon(True)
         th3.start()
+        th4 = threading.Thread(target = self.act_dome)
+        th4.setDaemon(True)
+        th4.start()
+        th5 = threading.Thread(target = self.stop_dome)
+        th5.setDaemon(True)
+        th5.start()
+
+
 
     ###set encoder az for dome tracking
     def set_enc_parameter(self, req):
@@ -477,79 +190,135 @@ class dome_controller(object):
         name = req.name
         value = req.value
         self.parameters[name] = value
-        self.flag = 0
+        self.paralist.append(self.parameters[name])
         print(name,value)
         return
 
+    def set_az_command(self, req):
+        self.parameter_az = req.value
+        return
+
     ###function call to dome/memb action 
-    def act_dome(self):
-        while True:
-            if self.flag == 1:
+    def dome_OC(self):
+        while not rospy.is_shutdown():
+            if self.paralist == []:
                 time.sleep(1)
                 continue
-            if self.parameters['command'] == 'pass':
+            elif "dome_open" in self.paralist and "dome_close" in self.paralist:
+                if self.paralist.index("dome_open") < self.paralist.index("dome_close"):
+                    self.dev.dome_open()
+                    self.paralist.remove("dome_open")
+                else:
+                    self.dev.dome_close()
+                    self.paralist.remove("dome_close")
+            elif "dome_open" in self.paralist:
+                self.dev.dome_open()
+                self.paralist.remove("dome_open")
+            elif "dome_close" in self.paralist:
+                self.dev.dome_close()
+                self.paralist.remove("dome_close")
+            else:
                 time.sleep(1)
-            elif self.parameters['command'] == 'dome_open':
-                self.dome_open()
-                self.flag = 1
-            elif self.parameters['command'] == 'dome_close':
-                self.dome_close()
-                self.flag = 1
-            elif self.parameters['command'] == 'memb_open':
-                self.memb_open()
-                self.flag = 1
-            elif self.parameters['command'] == 'memb_close':
-                self.memb_close()
-                self.flag = 1
-            elif self.parameters['command'] == 'dome_move':
-                self.move(self.parameters['target_az'])
-                self.flag = 1
-            elif self.parameters['command'] == 'dome_stop':
-                self.dome_stop()
-                self.end_flag = True
-                self.flag = 1
-            elif self.parameters['command'] == 'dome_tracking':
+                continue
+            time.sleep(1)
+            continue
+
+    def memb_OC(self):
+        while not rospy.is_shutdown():
+            if self.paralist == []:
+                time.sleep(1)
+                continue
+            elif "memb_open" in self.paralist and "memb_close" in self.paralist:
+                if self.paralist.index("memb_open") < self.paralist.index("memb_close"):
+                    self.dev.memb_open()
+                    self.paralist.remove("memb_open")
+                else:
+                    self.dev.memb_close()
+                    self.paralist.remove("memb_close")
+            elif "memb_open" in self.paralist:
+                self.dev.memb_open()
+                self.paralist.remove("memb_open")
+            elif "memb_close" in self.paralist:
+                self.dev.memb_close()
+                self.paralist.remove("memb_close")
+            else:
+                time.sleep(1)
+                continue
+            time.sleep(1)
+            continue
+
+    def act_dome(self):
+        while not rospy.is_shutdown():
+            if self.paralist == []:
+                time.sleep(1)
+                continue
+            elif "dome_move" in self.paralist and "dome_tracking" in self.paralist:
+                if self.paralist.index("dome_move") < self.paralist.index("dome_tracking"):
+                    sub3 = rospy.Subscriber('dome_move_az', Dome_msg, self.set_az_command)
+                    time.sleep(0.1)
+                    self.end_flag = False
+                    self.con_move(self.parameter_az)
+                else:
+                    self.end_flag = False
+                    self.con_move_track()
+            elif "dome_move" in self.paralist:
+                sub3 = rospy.Subscirber('dome_move_az', Dome_msg, self.set_az_command)
+                time.sleep(0.1)
                 self.end_flag = False
-                self.move_track()
-                self.flag = 1
-                pass
+                self.con_move(self.parameter_az)
+            elif "dome_tracking" in self.paralist:
+                self.end_flag = False
+                self.con_move_track()
+            else:
+                time.sleep(1)
+                continue
             time.sleep(1)
             continue
 
     def stop_dome(self):
-        while True:
-            if self.flag == 1:
-                time.sleep(1)
-                continue
-            elif self.parameters['command'] == 'dome_stop':
-                self.dome_stop()
+        while not rospy.is_shutdown():
+            if "dome_stop" in self.paralist:
+                self.dev.dome_stop()
                 self.end_flag = True
-                self.flag = 1
                 print('!!!dome_stop!!!')
-            elif self.parameters['command'] == 'dome_track_end':
+                self.paralist.remove("dome_stop")
+            elif "dome_track_end" in self.paralist:
                 self.end_flag = True
-                self.flag = 1
-                print('dome track end')
+                print("dome track end")
+                self.paralist.remove("dome_track_end")
+            elif "pass" in self.paralist:
+                time.sleep(1)
+                self.paralist.remove("pass")
+            else:
+                pass
             time.sleep(1)
-            continue        
+            continue
 
     ###publish status
+    @deco
     def pub_status(self):
-        pub = rospy.Publisher('status_dome', Status_dome_msg, queue_size=10, latch = True)
-        s = Status_dome_msg()
-        while True:
-            s.move_status = self.move_status
-            s.right_act = self.right_act
-            s.right_pos = self.right_pos
-            s.left_act = self.left_act
-            s.left_pos = self.left_pos
-            s.memb_act = self.memb_act
-            s.memb_pos = self.memb_pos
-            s.remote_status = self.remote_status
+        track_pub = rospy.Publisher('dome_track_flag', Bool_necst, queue_size=1)
+        while not rospy.is_shutdown():
+            pub = rospy.Publisher('status_dome', Status_dome_msg, queue_size=10, latch = True)
+            s = Status_dome_msg()
+            s.move_status = self.dev.move_status
+            s.right_act = self.dev.right_act
+            s.right_pos = self.dev.right_pos
+            s.left_act = self.dev.left_act
+            s.left_pos = self.dev.left_pos
+            s.memb_act = self.dev.memb_act
+            s.memb_pos = self.dev.memb_pos
+            s.remote_status = self.dev.remote_status
             s.dome_enc = float(self.dome_enc)
             s.from_node = node_name
             s.timestamp = time.time()
             pub.publish(s)
+            if self.end_flag:
+                track_pub.publish(False, node_name, time.time())
+            elif not self.end_flag:
+                track_pub.publish(True, node_name, time.time())
+            else:
+                print("where track flag???")
             time.sleep(0.1)
         
 if __name__ == '__main__':
